Improve documentation
Added examples. Named parameters. Added elading :: for references to SketchUp's own classes (otherwise links were formed to this library's corresponding classes).

Problems:

Still links are formed to library's classes in code examples. Code example's should not be run insside of the library's namespace and leading :: is therefore redundant. Is there a YARD setting for what namespace example code is in?

If a Sketchup module is added to the library, leading :: will also be needed in all references to the API's Sketchup module.