Bake in last matrix element in axis lengths
Makes *axis, scale*, determinant and flipped? all honour the scaling set in the last component of the matrix. This also makes scaling(-1, -1, -1) and scaling(-1) be regarded as the same Transformation, even on 2017 and lower versions.

Testing shows the last element is the inverse of the uniform scaling, not the uniform scaling as I though and wrote the previous code for.

SU2018 has changed not to use this component, but set distribute uniform scaling over all the axes instead.