Trying to make Geom one united Skippy Module
This solves the circular dependencies thing while still allowing the individual [Ruby] modules to refer to rely on each other.

However it forces any user who just wan't a potion of the functionality, e.g. testing what side on a plane a point is, to load the whole Geom package.