Strike over Skippy in Readme for now
ThomThom has asked me to not attract attention to Skippy as it is not ready for release. cmty-lib isn't planned to be released untilSkippy is, but in case people find it prior to that they should be warned about using Skippy.

References to Skippy can't be removed completely (without dropping support) until Skippy supports custom library module names (github/skippy#19). If implemented, this project could hide references to Skippy and be released without outing Skippy. Otherwise wait for Skippy to be released, and then revert this commit.